refac: move organizers to specific folder
