+++ spec/services/organizer/lecture_organizer_service_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
-require_relative '../fixtures/lectures_data'
+require_relative '../../fixtures/lectures_data'
 
-RSpec.describe LectureOrganizerBacktrackService, type: :service do
+RSpec.describe Organizer::LectureOrganizerService, type: :service do
   let!(:lectures_input) { LECTURE_DATA.dup }
-  let!(:expected_schedule_output) { File.read(Rails.root.join('spec/fixtures/expected_schedule_output.txt')) }
+  let!(:expected_schedule_output) { File.read(Rails.root.join('spec/fixtures/output_example.txt')) }
 
   describe '.organize' do
     context 'when there are no lectures to organize' do
@@ -22,10 +22,10 @@ RSpec.describe LectureOrganizerBacktrackService, type: :service do
       end
     end
 
-    context 'when not all lectures can be allocated within the day' do
+    context 'when not all lectures can be allocated within the tracks' do
       let!(:conference) { create(:conference) }
       let!(:track) { create(:track, conference: conference) }
-      let(:lectures) { create_list(:lecture, 99, track: track) }
+      let!(:lectures) { create_list(:lecture, 99, track: track) }
 
       it 'returns a message indicating not all lectures could be allocated' do
         result = described_class.organize(lectures)
